Bump version and fix text background once and for all

+++ src/main/java/io/github/ngspace/hudder/uielements/TextElement.java
@@ -10,13 +10,13 @@ public class TextElement extends AUIElement {
 	public final int x;
 	public final int y;
 	public final int color;
-	public final long backgroundcolor;
+	public final int backgroundcolor;
 	public final float scale;
 	public final boolean shadow;
 	public final boolean background;
 
-	public TextElement(int x, int y, String text, float scale, int color, boolean shadow,boolean background,
-			long backgroundcolor) {
+	public TextElement(int x, int y, String text, float scale, int color, boolean shadow, boolean background,
+			int backgroundcolor) {
 		this.text = text;
 		this.x = x;
 		this.y = y;
